+++ src/steps/resource-manager/storage/client.ts
@@ -3,6 +3,7 @@ import {
   BlobContainer,
   FileShare,
   StorageAccount,
+  Table,
   StorageQueue,
   Kind,
 } from '@azure/arm-storage/esm/models';
@@ -122,6 +123,30 @@ export class StorageClient extends Client {
         resourceGroupName: resourceGroup,
         accountName,
         resourceDescription: 'storage.queues',
+        callback,
+      });
+    }
+  }
+
+  /* eslint-disable @typescript-eslint/no-non-null-assertion */
+  public async iterateTables(
+    storageAccount: { name: string; id: string; kind: Kind },
+    callback: (e: Table) => void | Promise<void>,
+  ): Promise<void> {
+    const serviceClient = await this.getAuthenticatedServiceClient(
+      StorageManagementClient,
+    );
+    const resourceGroup = resourceGroupName(storageAccount.id, true)!;
+    const accountName = storageAccount.name!;
+
+    if ((['Storage', 'StorageV2'] as Kind[]).includes(storageAccount.kind)) {
+      return iterateAllStorageAccountResources({
+        logger: this.logger,
+        serviceClient,
+        resourceEndpoint: serviceClient.table,
+        resourceGroupName: resourceGroup,
+        accountName,
+        resourceDescription: 'storage.tables',
         callback,
       });
     }