+++ src/steps/resource-manager/storage/converters.ts
@@ -3,6 +3,7 @@ import {
   FileShare,
   StorageAccount,
   StorageQueue,
+  Table,
 } from '@azure/arm-storage/esm/models';
 import {
   createIntegrationEntity,
@@ -16,6 +17,7 @@ import {
   STORAGE_CONTAINER_ENTITY_METADATA,
   STORAGE_FILE_SHARE_ENTITY_METADATA,
   STORAGE_QUEUE_ENTITY_METADATA,
+  STORAGE_TABLE_ENTITY_METADATA,
 } from './constants';
 
 /**
@@ -144,6 +146,30 @@ export function createStorageQueueEntity(
         type: data.type,
         resourceGroup: resourceGroupName(data.id),
         encrypted: storageAccountEntity.encryptedQueue,
+      },
+    },
+  });
+}
+
+export function createStorageTableEntity(
+  webLinker: AzureWebLinker,
+  storageAccountEntity: Entity,
+  data: Table,
+): Entity {
+  return createIntegrationEntity({
+    entityData: {
+      source: data,
+      assign: {
+        _type: STORAGE_TABLE_ENTITY_METADATA._type,
+        _class: STORAGE_TABLE_ENTITY_METADATA._class,
+        webLink: webLinker.portalResourceUrl(data.id),
+        name: data.name,
+        displayName: data.name,
+        tableName: data.tableName,
+        type: data.type,
+        resourceGroup: resourceGroupName(data.id),
+        classification: null,
+        encrypted: !!storageAccountEntity.encryptedTable,
       },
     },
   });